+++ console_gpt/menus/assistant_menu.py
@@ -8,16 +8,15 @@ from typing import List, Optional, Tuple
 import openai
 import requests
 
-from console_gpt.config_manager import ASSISTANTS_PATH, fetch_variable, write_to_config
+from console_gpt.config_manager import (ASSISTANTS_PATH, fetch_variable,
+                                        write_to_config)
 from console_gpt.custom_stdin import custom_input
 from console_gpt.custom_stdout import custom_print
 from console_gpt.general_utils import capitalize, decapitalize
 from console_gpt.menus.role_menu import _add_custom_role, role_menu
-from console_gpt.menus.skeleton_menus import (
-    base_checkbox_menu,
-    base_multiselect_menu,
-    base_settings_menu,
-)
+from console_gpt.menus.skeleton_menus import (base_checkbox_menu,
+                                              base_multiselect_menu,
+                                              base_settings_menu)
 from console_gpt.prompts.file_prompt import _validate_file, browser_files
 from console_gpt.prompts.save_chat_prompt import _validate_confirmation
 from console_gpt.prompts.system_prompt import system_reply
@@ -151,6 +150,7 @@ def _selected_remote_files(model):
     else:
         return None
 
+
 def _upload_files(model):
     files = []
     while True:
@@ -165,6 +165,7 @@ def _upload_files(model):
             continue
         else:
             return files
+
 
 def _select_assistant_tools():
     tools_selection = base_settings_menu(
@@ -469,13 +470,15 @@ def _remove_assistant_files(model, assistant):
         assistant_fileurl = f"{OPENAI_URL}{ASSISTANTS_ENDPOINT}/{assistant_id}/files/{fileid}"
         response = requests.delete(assistant_fileurl, headers=assistant_headers).json()
         if response["deleted"] == True:
-            custom_print("info",
-                f'File "{next(item["filename"] for item in remote_assistant_files if item["id"]==fileid)}"" successfully removed from {assistant_name}.'
+            custom_print(
+                "info",
+                f'File "{next(item["filename"] for item in remote_assistant_files if item["id"]==fileid)}"" successfully removed from {assistant_name}.',
             )
         confirmation = _delete_file(model, fileid)
         if confirmation == True:
-            custom_print("info",
-                f'File "{next(item["filename"] for item in remote_assistant_files if item["id"]==fileid)}" deleted successfully.'
+            custom_print(
+                "info",
+                f'File "{next(item["filename"] for item in remote_assistant_files if item["id"]==fileid)}" deleted successfully.',
             )